+++ TR5Main/Objects/TR1/Entity/tr1_bear.cpp
@@ -56,8 +56,8 @@ void BearControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short head = 0;
-	short angle = 0;
+	float head = 0;
+	float angle = 0;
 
 	if (item->HitPoints <= 0)
 	{
