+++ TR5Main/Objects/TR2/Entity/tr2_shark.cpp
@@ -21,8 +21,8 @@ void SharkControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* info = GetCreatureInfo(item);
 
-	short angle = 0;
-	short head = 0;
+	float angle = 0;
+	float head = 0;
 
 	if (item->HitPoints <= 0)
 	{
